8th commit: i was able to resolve the error"transient error" that I mentioned earlier. So the OneToOne for both vehicle and customer worked after testing with the postman and the data was saved properly in the tables.

+++ VehicleProject/src/main/java/com/chika/service/CustServ.java
@@ -3,23 +3,28 @@ package com.chika.service;
 import java.util.ArrayList;
 import java.util.List;
 
+
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 
 import com.chika.model.Customer;
 
 import com.chika.repository.CustRep;
 
 @Service
+
 public class CustServ {
 
 	 @Autowired
 	    private CustRep custRepo;
 	        
-	    	public Customer getCustomerById(Integer custId) {
+	    	public Customer getCustomerById(Long custId) {
 	    		Customer custo= custRepo.findById(custId).get();
 	    		return custo;
 	    	}
+	    	
 
 	    	public List<Customer> getAllCustomers(){
 	    		List<Customer> list = new ArrayList<>();
